+++ src/bfabric_scripts/cli/cli_external_job.py
@@ -8,6 +8,7 @@ from rich.pretty import pprint
 from bfabric import Bfabric
 from bfabric.wrapper_creator.bfabric_submitter import BfabricSubmitter
 from bfabric.wrapper_creator.bfabric_wrapper_creator import BfabricWrapperCreator
+from bfabric_scripts.cli.external_job.upload_submitter_executable import upload_submitter_executable_impl
 from bfabric_scripts.cli.external_job.upload_wrapper_creator_executable import upload_wrapper_creator_executable_impl
 
 app = cyclopts.App()
@@ -38,6 +39,20 @@ def wrapper_creator(external_job_id: int) -> None:
     client = Bfabric.from_config()
     creator = BfabricWrapperCreator(client=client, external_job_id=external_job_id)
     pprint(creator.create())
+
+
+@app.command
+def upload_submitter_executable(
+    filename: Path,
+    *,
+    engine: Literal["slurm"] = "slurm",
+    name: str = "yaml / Slurm executable",
+    description: str = "Submitter executable for the bfabric functional test using Slurm.",
+) -> None:
+    client = Bfabric.from_config()
+    upload_submitter_executable_impl(
+        client=client, filename=filename, engine=engine, name=name, description=description
+    )
 
 
 @app.command